+++ django_typer/management/__init__.py
@@ -32,9 +32,9 @@ from typer.main import get_params_convertors_ctx_param_name_from_function  # noq
 from typer.models import Context as TyperContext  # noqa: E402
 from typer.models import Default, DefaultPlaceholder  # noqa: E402
 
-from ..completers import ModelObjectCompleter  # noqa: E402
+from ..completers.model import ModelObjectCompleter  # noqa: E402
 from ..config import traceback_config  # noqa: E402
-from ..parsers import ModelObjectParser  # noqa: E402
+from ..parsers.model import ModelObjectParser  # noqa: E402
 from ..types import (  # noqa: E402
     ForceColor,
     NoColor,