+++ examples/polls/management/commands/closepoll_t5.py
@@ -4,8 +4,8 @@ from django.utils.translation import gettext_lazy as _
 from typer import Argument, Option
 
 from django_typer.management import TyperCommand
-from django_typer.completers import ModelObjectCompleter
-from django_typer.parsers import ModelObjectParser
+from django_typer.completers.model import ModelObjectCompleter
+from django_typer.parsers.model import ModelObjectParser
 from tests.apps.examples.polls.models import Question as Poll