+++ tests/apps/perf_typer/management/commands/perf.py
@@ -1,4 +1,5 @@
 from django_typer.management import TyperCommand
+import json
 import sys
 
 
@@ -7,4 +8,10 @@ class Command(TyperCommand):
     requires_migrations_checks = False
 
     def handle(self, test_arg: int, test_option: bool = False):
-        return f"typer: {test_arg}, test_option={test_option}, {len(sys.modules)}"
+        return json.dumps(
+            {
+                "typer": test_arg,
+                "test_option": test_option,
+                "modules": list(sys.modules.keys()),
+            }
+        )