+++ tests/apps/test_app/management/commands/model_fields.py
@@ -17,7 +17,7 @@ from django_typer.management import (
     model_parser_completer,
 )
 from django_typer import types
-from django_typer.completers import ModelObjectCompleter
+from django_typer.completers.model import ModelObjectCompleter
 from tests.apps.test_app.models import ShellCompleteTester